API/requirements.txt
bs4
fastapi
pydantic
httpx
uvicorn
google-generativeai
python-dotenv